# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zytrixon Mobile API Documentation",
        "description": "API documentation for the Zytrixon Truck Billing mobile application.",
        "contact": {
            "email": "admin@zytrixon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get all vehicles",
                "operationId": "028ac63ffd1730fb647999d1bc6e77d9",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicles retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Create a new vehicle",
                "operationId": "a396cd54d88d3edeb8c6488411c67eb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "truck_number",
                                    "make",
                                    "model",
                                    "year",
                                    "chassis_number",
                                    "engine_number"
                                ],
                                "properties": {
                                    "truck_number": {
                                        "type": "string",
                                        "example": "MH-12-AB-1234"
                                    },
                                    "make": {
                                        "type": "string",
                                        "example": "Tata"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Signa"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2021
                                    },
                                    "chassis_number": {
                                        "type": "string",
                                        "example": "CH123456789"
                                    },
                                    "engine_number": {
                                        "type": "string",
                                        "example": "EN123456789"
                                    },
                                    "rc_document": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicles/stats": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get vehicle statistics",
                "operationId": "142da73bab186990da0df95243a19362",
                "responses": {
                    "200": {
                        "description": "Stats retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get specific vehicle",
                "operationId": "82267179ee94cba8d4d423f4265030c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehicle retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wallets": {
            "get": {
                "tags": [
                    "Driver Wallets"
                ],
                "summary": "Get driver wallets",
                "description": "Owner gets all driver wallets, driver gets their own wallet.",
                "operationId": "6a93b0c63ef54f291bb0aa171e7cfe3e",
                "responses": {
                    "200": {
                        "description": "Wallets retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wallets/add-funds": {
            "post": {
                "tags": [
                    "Driver Wallets"
                ],
                "summary": "Add funds to a driver wallet",
                "operationId": "6d2d2d45ded68888dc9b03903fdca780",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "driver_id",
                                    "amount"
                                ],
                                "properties": {
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 1000
                                    },
                                    "remarks": {
                                        "type": "string",
                                        "example": "Advance pay",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funds added successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Sanctum token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}